+++ src/App.jsx
@@ -37,8 +37,8 @@ import Finance from "./Finance";
 import Budgeting from "./Budgeting";
 import Goals from "./Goals";
 
-// Mental Wellness module - Saeed Peyman
-import Mental from "./Mental";
+// Mental Health Chatbot module
+import Mental from "./MentalHealthChatbot";
               
 // Habit Tracker module 
 import HabitTracker from "./HabitTracker";
@@ -143,7 +143,7 @@ class App extends Component {
                   <NavLink to="/finance">Finance</NavLink>
                 </li>
 
-                {/* Mental Wellness */}
+                {/* Mental Health Chatbot */}
                 <li>
                   <NavLink to="/mental">Mental Health Chatbot</NavLink>
                 </li>
